+++ internal/build-index.sh
@@ -7,14 +7,8 @@ shift
 
 TEMP=$(mktemp)
 
-cat << EOF > "${TEMP}"
-m4_include(include/html.m4)
-START_HTML([[${ROOT_URL}]], [[${BLOG_TITLE} - Home]])
-HOMEPAGE_HEADER_HTML([[${ROOT_URL}]], [[${BLOG_TITLE}]], [[${BLOG_SUBTITLE}]])
-EOF
-
-while read FILE
-do
+function do_file {
+	FILE="$1"
 	POST_LINK="${ROOT_URL}/posts/$(basename "${FILE}" ".${POST_EXTENSION}").html"
 	TITLE="$(get_title "${FILE}")"
 	DATE="$(get_date "${FILE}")"
@@ -27,6 +21,10 @@ do
 
 	CONTENT_IS_TRIMMED="$(content_will_be_trimmed "${FILE}")"
 
+	OPTIONS="$(parse_options "${FILE}")"
+	IS_PINNED="$(op_get "${OPTIONS}" pinned)"
+	rm "${OPTIONS}"
+
 	CONTENT="$(get_and_parse_content "${FILE}" "trimmed")"
 	cat << EOF >> "${TEMP}"
 START_HOMEPAGE_PREVIEW_HTML
@@ -38,8 +36,10 @@ EOF
 POST_HEADER_MOD_DATE_HTML([[${MOD_DATE}]])
 EOF
 	fi
+	if [[ "${IS_PINNED}" != "" ]] && (( "${IS_PINNED}" > "0" ))
+	then
 		cat << EOF >> "${TEMP}"
-POST_HEADER_HTML([[<a href='${POST_LINK}'>${TITLE}</a>]], [[${DATE}]], [[${AUTHOR}]])
+POST_HEADER_PINNED_HTML
 EOF
 	fi
 	cat << EOF >> "${TEMP}"
@@ -55,8 +55,33 @@ EOF
 	cat << EOF >> "${TEMP}"
 END_HOMEPAGE_PREVIEW_HTML
 EOF
-	shift
-done < <(sort_by_date "$@" | tac | head -n "${POSTS_ON_HOMEPAGE}")
+}
+
+cat << EOF > "${TEMP}"
+m4_include(include/html.m4)
+START_HTML([[${ROOT_URL}]], [[${BLOG_TITLE} - Home]])
+HOMEPAGE_HEADER_HTML([[${ROOT_URL}]], [[${BLOG_TITLE}]], [[${BLOG_SUBTITLE}]])
+EOF
+
+IGNORE_POSTS=( )
+NUM_IGNORE_POSTS="0"
+
+while read FILE
+do
+	do_file "${FILE}"
+	IGNORE_POSTS["${NUM_IGNORE_POSTS}"]="${FILE}"
+	NUM_IGNORE_POSTS=$((NUM_IGNORE_POSTS+1))
+done < <(only_pinned_posts "$@")
+
+POSTS_REMAINING="${POSTS_ON_HOMEPAGE}"
+
+while read FILE
+do
+	[[ " ${IGNORE_POSTS[@]} " =~ " ${FILE} " ]] && continue
+	do_file "${FILE}"
+	POSTS_REMAINING=$((POSTS_REMAINING-1))
+	[[ "${POSTS_REMAINING}" == "0" ]] && break
+done < <(sort_by_date "$@" | tac)
 
 cat << EOF >> "${TEMP}"
 HOMEPAGE_FOOTER_HTML([[${ROOT_URL}]], [[${VERSION}]])